telemetry_project/requirements.txt
Django
djangorestframework
django-cors-headers
pyserial